style: moved state logic and firebase into core/
- Moved contexts/ into core/
- Moved hooks/ into core/
- Moved providers/ into core/
- Moved firebase.ts into core/api/
