+++ src/components/ThemeToggleButton.tsx
@@ -1,6 +1,6 @@
 import { IconButton } from "@mui/material"
 import { Brightness4, Brightness7 } from "@mui/icons-material"
-import { useThemeContext } from "@/hooks/useThemeContext"
+import { useThemeContext } from "@/core/hooks/useThemeContext"
 
 const ThemeToggleButton = () => {
   const theme = useThemeContext()
