+++ src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from "react"
 import { createRoot } from "react-dom/client"
-import App from "./App.tsx"
-import "./style/index.css"
-import AppProviders from "./providers/AppProviders.tsx"
+import App from "./App"
+import AppProviders from "./core/providers/AppProviders"
+import "./styles/index.css"
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
